First tranch of work to get W3SIN4 parallelised. Current progress:
L447 - Why does CREATING EB, EBX and EBY on the GPU cause results to be incorrect?
L453 - Explicity stating LOOP REDUCION of UORB and AORB causes results to be incorrect
L460 - Loop is sequential. Why? I've said it is independent
L554 - This double loop is a pain in the arse. I've had to make the following changes to get the inner loop to vercorize (outer loops still has a dependence on DSTAB):
STRESSTAB and STRESSTABN are two-d arrays and are used in a erduction operation. ACC didn't like this. However, they are of size (3,2) and only the 3rd element of the first dimension is ever used, so I have simply turned each into two scalar variables instead.
The LAST values of COSU and SINU calculated at L562 are used outside the loop. I've had to add a repeat calculation at L665 to get around this.

Next thing to do - get DSTAB entirely on the GPU (from line 672) as it is one of the larger arrays.